requirements.txt
Django~=5.2.7
pip~=24.3.1
cryptography~=41.0.4
filelock~=3.16.1
asgiref~=3.10.0
sqlparse~=0.5.3
Jinja2~=3.1.6
requests~=2.31.0
django-embed-video~=1.4.10